Added an endpoint where you can send the starting position in pi youd like to start grabbing digits from and it will return a 50x50 image.
